backend/requirements.txt

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
requests>=2.31.0
httpx>=0.26.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
redis>=5.0.1
celery>=5.3.6
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
facebook-sdk>=3.1.0  # For basic graph API interaction
